#38 moves validate module to lib
